do not keep tha mash_dist for each alignment (if not used)

+++ src/common/wflign/src/rkmh.hpp
@@ -47,6 +47,6 @@ std::vector<hash_t> hash_sequence(const char* seq,
                                   const uint64_t& k,
                                   const uint64_t& sketch_size);
 
-double compare(const std::vector<hash_t>& alpha, const std::vector<hash_t>& beta, const uint64_t& k);
+float compare(const std::vector<hash_t>& alpha, const std::vector<hash_t>& beta, const uint64_t& k);
 
 }
